Refactor create command - move GitRepository operations to create source git - update examples and e2e tests
