+++ .github/workflows/e2e.yaml
@@ -51,7 +51,9 @@ jobs:
           ./bin/tk install --manifests ./manifests/install/
       - name: Run create source e2e tests
         run: |
-          ./bin/tk create source podinfo --git-url https://github.com/stefanprodan/podinfo-deploy  --git-semver=">=0.0.1-rc.1 <0.1.0"
+          ./bin/tk create source git podinfo \
+            --git-url https://github.com/stefanprodan/podinfo-deploy  \
+            --git-semver=">=0.0.1-rc.1 <0.1.0"
       - name: Debug failure
         if: failure()
         run: |
